improvements for package qt  (by Mark Brand)
Added support for compiler flag rtti to configure script. This was
already supported by the configure binary for Windows.

Fixed "detection" of mmx, 3dnow, sse, sse2, and rtti in configure
script for win32-g++ target.

Fixed failure to add a whole bunch of CONFIG options to qconfig.pri,
including but not limited to mmx, 3dnow, sse, sse2. I wonder if this was
a bug in the configure script. Now the CONFIG line looks more like the
one produced by the configure binary.