do not introduce QtCore4 in ssl.pri
It is not correct to add QtCore4 in ssl.pri when -lz is removed
for qt-zlib.

In mingw-cross-env we add QtCore4 to OPENSSL_LIBS so the zlib
dependency of openssl is satisfied in case of qt-zlib. This
is necessary for the time being because the QtCore4 in the link
command that qmake generates appears before the contents of
OPENSSL_LIBS.