use correct options for arch and largefile for windows target of package qt
Previously we did not specify -arch, so the configure script used
the host architecture as the default. This worked as long as the
host and target architectures matched. That's why we had to force
it with -host-arch i386 to make it work on 64 bit hosts.

The correct value "windows" may have different properties than
i386, but this has not be researched. In any case "windows" is
the value use by the binary executable configure tool for windows.

Additionally, the configure script did not respect -arch when
cross-building on MAC. Now CFG_ARCH is set automatically to
windows to cover this case.

Also, "-largefile" is not appropriate for Windows. Now specify
"-no-largefile" to avoid defintions meant for Unix. Windows
apparently still has its largefile support.