Restore HTML compliance of index.html by declaring it as HTML5
Previously, index.html was declared as HTML 4.01 Strict.
However, HTML 4.01 doesn't allow for "+" characters in
IDs such as id="libodbc++-package". We could disallow
the "+" for all package names, but it seems to be easier
to just declare index.html as HTML5, as it apparently
allows for a wider range of characters in IDs, including "+".