update download rules from mxe
Previously, the download rules would only move on to a backup URL if
wget returned a failure status.  Now we will also move on to a backup
URL if wget succeeds but the checksum is incorrect.

Also look at an octave.org server as a backup location for package
sources.